Move shell command helpers out of harness
