+++ test/server_harness_test.go
@@ -1700,19 +1700,6 @@ func (h *ServerHarness) runShellCommand(pane, command, marker string) {
 	h.waitIdle(pane)
 }
 
-func shellSingleQuote(s string) string {
-	return "'" + strings.ReplaceAll(s, "'", `'"'"'`) + "'"
-}
-
-func nestedAmuxCommand(binPath, session string, args ...string) string {
-	parts := make([]string, 0, len(args)+3)
-	parts = append(parts, shellSingleQuote(binPath), "-s", shellSingleQuote(session))
-	for _, arg := range args {
-		parts = append(parts, shellSingleQuote(arg))
-	}
-	return strings.Join(parts, " ")
-}
-
 func (h *ServerHarness) sendClientKeys(keys ...string) string {
 	h.tb.Helper()
 	pane := h.activePaneName()
